- Data Feed is now functional interface instead of abstract class.
This is because the only behaviour a data feed should have
is to be able to be ingested into the database - the way it is
ingested is upto the data feed itself.

- Data feed manager now only needs to call the ingest method for every
data feed.

- ECB data feed has been updated to implement the ingest method, and
logic specific to the ECB data feed has been moved from the data feed
manager class into the ECB class.

- ECB data feed now used a HTTP client for HTTP requests and responses,
as this is a more standard/extensible way of downloading data from the
internet instead of using the URL class, input streams and channels.