- Changes to the beans created in the Spring Boot web app config class
Exchange Rate Service is a class only used by the web app, and so there is
no need to explicitly create a bean out of it in the config class, instead
 the Service annotation can be used to instantiate it.

Due to using a HTTP client in the ECB data feed, there is some additional
set-up required in the config class.

dataFeeds and dataFeedManager methods are no longer annotated with Bean
annotation as they are only used in the Rate Scheduler class, which is a
bean, and are not used/injected in any other component in this
web app, they have also been marked private to make this intention clearer.